pyproject.toml
[project]
name = "poc-lightspeed-agentic"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "llama-stack-client>=0.7.1",
    "llama-stack[starter]>=0.7.1",
    "pydantic-ai>=0.2.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
    "uvicorn>=0.34.0",
]
